requirements.txt
boto3==1.28.15
numpy==1.25.1
Pillow==10.0.0
qdrant_client==1.3.2
streamlit==1.25.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
